Added make parameters to makeall.sh.
Added make parameters to the makeall.sh script for building all
frontends.

+++ software/common/frontends/makeall.sh
@@ -9,14 +9,16 @@ if [ ! -d "./bin" ]; then
    mkdir bin
 fi
 
+make_params=$*
+
 # Attempt to 'make' in every directory not named "bin".
 # Then, attempt to copy a binary of the same name into the directory named "bin".
 for ii in `ls -1`; do
    if [ "bin" != $ii ]; then
       if [ -d $ii ]; then
          cd $ii
-         make
-         if [ 0 -eq $? ]; then
+         make $make_params
+         if [ 0 -eq $? ] && [ -e $ii ]; then
             chmod u+x $ii
             cp -pf $ii ../bin
          fi
